requirements.txt
openai
google-generativeai>=0.3.0
transformers>=4.36.0
torch>=2.0.0
huggingface_hub>=0.19.0
python-dotenv>=1.0.0
streamlit>=1.30.0
requests>=2.31.0
together>=0.2.0 